systems: Avoid redundantly computing subsystem feedthrough maps
The key observation is that HasDirectFeedthrough was always computing
the full GetDirectFeedthroughs and then throwing away all but the one
map entry.

We can compute the map once per system and reuse it as we search,
instead of once per possible input.  This removes some super-linear
growth factor in diagram size that ends up being problematic.