Deprecate SignedDistanceToPoint::is_grad_w_unique
The flag reporting whether the gradient is unique had several problems:
  1. It wasn't always valid (box and cylinder ignored questions of
     uniqueness when the query point was *inside*).
  2. That information alone is not particularly actionable; it
     would be better to know the full space of valid vectors.
  3. We have no code that actually makes use of this classifier
     (and no optimization code that even *can*).

Therefore, we're not computing it correctly, not using it, and
have no plans to do so, we'll simplify the code by deprecating and
then removing this flag.

The tests in proximitiy_engine_test are inherently redundant of the tests
for the callback directly. These test haven't be triaged and culled yet,
but for the purpose of this deprecation, it simply won't pay attention to
whether or not the gradient is unique or not.

NOTE: The python bindings for the query result never bound the
boolean flag, means no deprecation effort is necessary in the
bindings.