Push more of distance_to_shape_callback into .cc file
The limiting factor for pushing more of the code into the .cc file
was the logic for determining if calling the fallback was required.
That logic was buried in `ComputeNarrowPhaseDistance`. So, to
confirm that the logic was right, I specialized the infrastructure
on a scalar not otherwise supported by Drake (e.g., `float`). This
worked, but was unwieldy and required a fair amount of the code
to remain in the header (in order to instantiate on double).

The simple solution was to move the classification logic (fallback
required or not) into its own stand-alone function and test that
explicitly.

This led to a change in the unit tests to test this new function
directly. Part of this effort led to some test clean up:
  1. The logic for testing what is supported and what isn't needed
     to be updated to reflect the full set of primitives.
  2. Some unused code infrastructure (ComputeNarrowHalfspaceTest)
     was detected and removed.