WIP: Fix various details.
Most of these are just the working through of name or type changes.

The `if constexpr' blocks in leaf_system.h allow the pybind layer to
continue to work with Event base types; they might be a target for a
later pass of heap reform.

Not dealt with in any way are whatever deprecation obligations we have
incurred; I haven't even tried to analyze what others we might need.