[geometry] Push `distance_to_shape_callback` implementation into .cc (#14794)
Historically, the limiting factor for pushing more of the code into
the .cc file was the logic for determining if calling the fallback
was required. That logic was buried in ComputeNarrowPhaseDistance.
Testing that logic required extraordinary contortions and required
the templated code to be fully available.

The simple solution was to move the classification logic (fallback
required or not) into its own stand-alone function and test that
explicitly.

Based on this change:
  - The unit test is updated vis a vis fallback classificaiton.
  - The state of supported/unsupported geometries has been brought
    up to reflect the current state of the code (e.g., capsules,
    half spaces.)