direct_transcription: Parallelize test cases
This commit replaces the first-aid fix of #14854, which addressed test
timeouts introduced at #14840.

Run the test in 2 shards. Most of the cases are vanishingly fast, but
the discrete- and continuous-mode SNOPT solves of Multibody*Test are
quite long under slow tooling (ASAN). This reshuffle should bring the
test latency down enough to avoid timeouts.

Also, a minor tweak: remove or re-spell some uses of DRAKE_DEMAND, to
make the test more idiomatic.