DecomposeLumpedParameters from VectorX<Expression>
First implementation of lumped parameter decomposition for full symbolic::Expressions.

This updates (and will soon replace) our related implementation in solvers/system_identification.h .  That implementation used the old (to be expunged) Polynomial class, and required a manual step to substitute trig functions into polynomial functions.  This substitution is difficult in general, and is not compatible with MultibodyPlant.  Furthermore, there is no fundamental reason to restrict the lumped-parameter decomposition to polynomials.

Follow-up work is tracked in #14879.