DecomposeLumpedParameters from VectorX<Expression>
First implementation of lumped parameter decomposition for full symbolic::Expressions.

This updates (and will soon replace) our related implementation in solvers/system_identification.h .  That implementation used the old (to be expunged) Polynomial class, and required a manual step to substitute trig functions into polynomial functions.  This substitution is difficult in general, and is not compatible with MultibodyPlant.  Furthermore, there is no fundamental reason to restrict the lumped-parameter decomposition to polynomials.

The follow-up work after this PR is to:
1) implement the numerical linear algebra logic for finding "identifiable lumped parameters"
2) provide a full system identification example using MultibodyPlant (i have a WIP of this in the sysid chapter of my underactuated notes using the cart pole).
3) deprecate and remove solvers/system_identification
4) implement simplification procedures for the Expression;  the general form is likely a `RowReduction` (e.g. Gaussian elimination) algorithm for a matrix of Expressions that uses a visitor that checks if two expressions are structurally equivalent up to a constant.  So, for example, `EqualToUpToConstantFactor(x, 3x)` returns `true` or `3`. But `EqualToUpToConstantFactor(x, 3x + 2)` returns `false`.