Deprecation - All things geometry (#14855)
While this deprecation effort contains the normal deletion of deprecated
API and its tests, it comes with a number of special cases -- some
transitionary code had to be injected during the deprecation period between
the old and new APIs. The explicit examples are called out below:

 - RenderEngine
   - The old API and new API were entangled such that derived RenderEngine
     types could overload either API and invoking one would eventually
     invoke the other. That mechanism has been removed and the default
     implementation now simply throws "not implemented" exceptions.
 - SceneGraph
   - SceneGraph went from storing GeometryState as abstract state to as an
     abstract Parameter. During transition, one could request the old
     behavior. That facility is now removed as well as the infrastructure
     that enabled it.

Implicit is the removal of *two* headers (but the contents of those
headers had already been completely deprecated).