Add template specialization of not_equal_strict for symbolic::Expression (#14868)
This is needed for Eigen >= 3.4.0. not_equal_strict relies on an
implicit conversion to bool. Since Expression has an explicit bool()
operator, we need to specialize the call. As of Eigen 3.4.0, this
function is called at least as part of triangular vector solve.
not_equal_strict was only added in Eigen 3.3.5. Since the minimum
version drake requires is 3.3.4, we have to conditionally compile.