[geometry] Deprecate the "is grad unique" flag from query results (#14875)
The flag reporting whether the gradient is unique had several problems:
  1. It wasn't always valid (box and cylinder ignored questions of
     uniqueness when the query point was *inside*).
  2. That information alone is not particularly actionable; it
     would be better to know the full space of valid vectors.
  3. We have no code that actually makes use of this classifier
     (and no optimization code that even *can*).

Therefore, we're not computing it correctly, not using it, and
have no plans to do so, we'll simplify the code by deprecating and
then removing this flag.

The tests in proximitiy_engine_test are inherently redundant of the tests
for the callback directly. These test haven't be triaged and culled yet,
but for the purpose of this deprecation, it simply won't pay attention to
whether or not the gradient is unique or not.

NOTE: The python bindings for the query result never bound the
boolean flag, means no deprecation effort is necessary in the
bindings.

* Deprecating vector is unique in SignedDistancePair

Deprecate SignedDistancePair::is_nhat_BA_W_unique

This follows logically from the deprecation of
SignedDistanceToPoint::is_grad_W_unique.

There are a couple of differences compared to the
SigndDistanceToPoint deprecation.
 - is_nhat_BA_W_unique was bound in python bindings, so this
   includes deprecations in the bindings as well.
 - SignedDistancePair had a weird constructor that only existed
   for tests. Rather than trying to maintain a version into the
   future, the tests were reformulated so that the constructor
   is no longer called and it can simply be deprecated in
   anticipation of simple removal.