[proximity] Eliminate MeshField API (with breaking changes) (#14897)
The MeshField API was speculative -- the idea that we'd represent the
mesh field multiple ways. This has not proven to be the case and the base
class is now an impediment. It will become more apparent in the future, but
essentially, the virtual API on MeshField interferes with differentiable
contact. So, we're eliminating the API and simply rolling the limited
amount of base class API into its only implementation: MeshFieldLinear.

This also leads to some minor administrative clean up.

 - mesh_field_linear.h defined an alias for SurfaceMeshFieldLinear but
   VolumeMeshFieldLinear was defined in its own header file. We've created
   surface_mesh_field.h for symmetric treatment.
 - The pre-built versions of {Volume|Surface}MeshFieldLinear based on
   double and AutoDiffXd are now distributed to the mesh-valued files
   *_mesh_field.{h,cc}. This required complete definition of MeshField in
   the header.

This introduces a number of breaking changes:

 - Deletion of mesh_field.{h|cc} and the class (in the drake:geometry
   namespace).
 - ContactSurface changes its type of field() from SurfaceMeshField to
   SurfaceMeshFieldLinear
 - Code that was getting SurfaceMeshFieldLinear from mesh_field_linear.h
   must now get it from surface_mesh_field.h

However, given that these are all components of hydroelastic which is
clearly labeled as being "under construction", these API breakages are
acceptable (particularly compared to the cost of attempting to deprecate).