framework: Avoid functor allocation for all event types (breaking change)
Relevant to: #14543

This is the seventh of a long PR train to make heapless simulation
possible, with careful system construction. Inspired by @edrumwri's
draft PR #14707.

This patch extends the techniques of PR #14969 to all of the event
types.

Subsequent patches will address other sources of heap allocation.

This is a breaking change because it removes the two-argument handle()
methods and replaces them with a three-argument form. The known uses
inside Drake are inside the system framework, or in tests. All of those
are updated. There are no known uses outside of Drake.