workspace: Upgrade libsdformat to latest release 11.0.0
This is a breaking change in the sense that this release provides stricter
validation of SDFormat  input data, so bad files that used to parse without
complaint now generate errors.  For example, the `<pose/>` element now has
more syntax checking that previously.

This is a breaking change in the sense that `//include` tags are presently
disabled in this version of Drake. In a subsequent PR, this functionality
will be restored with changes that will be detailed upon in that PR.

If you encounter issues with `//include` as they are used in Drake in this
interim period, please notify Eric Cousineau.

Co-authored-by: "Addisu Z. Taddese" <addisu@openrobotics.org>
Co-authored-by: "Eric Cousineau" <eric.cousineau@tri.global>
Co-authored-by: "Sean Curtis" <sean.curtis@tri.global>