workspace: Upgrade libsdformat to latest release 11.0.0
This is a breaking change in that <include> elements are no longer
supported.  In the future, this feature will be restored with revised
semantics.  If you encounter issues with <include> elements in the
interim, please notify Eric Cousineau.

This is a breaking change in the sense that this release provides stricter
validation of SDFormat input data, so bad files that used to parse without
complaint now generate errors.  For example, the <pose> element now has
more syntax checking than previously.

Co-authored-by: "Addisu Z. Taddese" <addisu@openrobotics.org>
Co-authored-by: "Eric Cousineau" <eric.cousineau@tri.global>
Co-authored-by: "Sean Curtis" <sean.curtis@tri.global>