sensors: Fix image_to_lcm_image_array uninitialized data
The implementation was leaving the outer msg->header.seq unset.

The implementation strategy was also unclear as to why we were taking
the long way around.  I suspect it was for efficiency's sake, so here
we document that fact along with comments to explain its correctness,
as well as removing a large inner loop heap use + redundant copy when
compression was enabled.

In the header, also shuffle the constructor to the top of the class
per the style guide, and also improve its API doc for clarity.

Disclaim a TODO to set the image sequence number.  Doing so would be
an odd job for a feedthrough system.  Assigning a sequence number
typically requires a clock.  This system should just leave it at zero.