New method for forming a rotation matrix from a single unit vector. (#14816)
Provide two new static methods on RotationMatrix to create
a RotationMatrix (i.e., basis) from a single vector. This 
should eventually replace the function in 
orthonormal_basis.h as this is more robust and faster.

The two variants are the user-friendly "we'll normalize the
vector for you" and the advanced "you're sure the vector is
already normalized?" variants.