Add DiscreteUpdateManager and PhysicalModel
Add a DiscreteUpdateManager class that advances the discrete states for
MultibodyPlant. DiscreteUpdateManager provides an interface that
extends MbP and that MbP knows how to invoke. It allows more
flexibility in experimenting with new solvers and can be compiled
independently from MbP.

Add a PhysicalModel class that extends the physical models owned by
MultibodyPlant. It allows more flexibility in experimenting new
materials and models that may be simulated.

To enable DiscreteUpdateManager and PhysicalModel, introduce attorney
classes MultibodyPlantModelAttorney and
MultibodyPlantDiscreteUpdateManagerAttorney that grant the necessary
protected/private methods in MultibodyPlant that PhysicalModel and
DiscreteUpdateManager needs.

Co-authored-by: amcastro-tri alejandro.castro@tri.global