framework: Avoid functor allocation for PublishEvents (#14969)
* framework: Avoid functor allocation for PublishEvents (breaking change)

Relevant to: #14543

This is the fifth of a long PR train to make heapless simulation
possible, with careful system construction. Inspired by @edrumwri's
draft PR #14707.

This patch introduces a new callback type for PublishEvents that avoids
heap allocations induced by lambda captures.

Subsequent patches will expand this technique (and test coverage) to the
other event types.

This is a breaking change because it removes the one-argument handle() method
and replaces it with a two-argument form. The known uses inside Drake are
inside the system framework, or in tests. All of those are updated. There are
no known uses outside of Drake.