primitives: Establish AffineSystem feedthrough without symbolics
The output port of AffineSystem now responds to feedthrough queries
without first scalar-converting itself to symbolic expressions.

This same change applies to LinearSystem, which inherits form
AffineSystem.

This is likely to be more performant, and also ensures operation even
for subclasses of affine or linear systems that choose not to support
symbolic scalar conversion.