Add is_convex flag to each QuadraticCost (#15043)
This PR doesn't check if the aggregated cost is convex or not. If we
have two quadratic costs cost1 and cost2, where cost1 is convex, cost2
is non-convex, and cost1 + cost2 is still convex. The convex solvers
can still solve this problem, but currently we will require each cost
to be convex, so that we can pass the MathematicalProgram to the
convex solver. I made this choice for two reasons:

i. It is computationally expensive/unreliable to determine if the
aggregated cost is convex or not. The Hessian of the aggregated cost
is likely to be a big sparse matrix, but Eigen doesn't provide a good
LDLT solver for sparse matrix. The available sparse LDLT solvers in
Eigen require LGPL license
https://eigen.tuxfamily.org/dox/group__TopicSparseSystems.html.

ii. This is the same behavior of CVX, which would throw an error if
any individual cost is non-convex, even though the summed-up cost is
convex.