Create general bounding volume hierarchy
The Bvh code is now templated on bounding volume type so the same
*hierarchy* logic can be used for different bounding volume types.
Furthermore, support collision *between* hierarchies of different bounding
volume types (in so far as the bounding volumes themselves support
overlap tests.

The majority of the meaningful changes are in:
  bvh.{h|cc}, obb.h and bvh_test.cc.

All other changes are simply respellings of the Bvh invocation to account
for the change in type declaration.