pydrake: Remove spurious documentation header dependencies
These were added in 6f37fd67 at a time when our pydrake linking was broken.
We were linking the examples' libraries outside of libdrake.so, so their
headers were not part of libdrake's headers, so for mkdoc to find them we
needed to list them here individually.

Now that we link the examples correctly (i.e., directly into libdrake.so),
we no longer need to feed the redundant dependencies to mkdoc.

As of bazel 4.1, the redundant headers actually break mkdoc, because we try
to add these headers twice, with the second path spelling not matching any
valid include path.