mkdoc: Work around a problematic deprecation_example header dep
When collecting the list of package headers, in the case of the deprecation
example (which has no package library) we were ending up with two different
copies of the header being passed to mkdoc:

  bazel-out/k8-opt/bin/bindings/pydrake/common/_virtual_includes/deprecation_example_class/drake/bindings/pydrake/common/test/deprecation_example/example_class.h
  bindings/pydrake/common/test/deprecation_example/example_class.h

The first one is the path we want -- the one used while compiling, and thus
the one with a matching -I flag.  The second one is a correct dependency,
but not used for compiling, so does not have any -I flag, so breaks mkdoc.

Here, we customize our skylark header collection to ignore the second file.