Adding collision filter manager API
In anticipation of expanding the functionality relating to working
with collision filters, we're changing the public API. This introduces
a *parallel* API and labels the old API as "legacy". The legacy API
will be deprecated in a follow up PR.

The purpose of the API is to reduce SceneGraph's knowledge of working
with collision filters. SceneGraph produces a CollisionFilterManager,
but all manipulation of collision filter configuration now belongs to
that manager (reducing the churn on SceneGraph's already large API).

To achieve this end:

  - Introduce new classes: CollisionFilterManager and
    CollisionFilterDeclaration.
    - The unit tests for CollisionFilterManager are copied from
      scene_graph_test.cc (as the manager can only be constructed by
      SceneGraph). The source test code is marked as Legacy for future
      deprecation.
  - Add new API: SceneGraph::collision_filter_manager().
  - GeometryState updates API to support new manager/declaration API.

This does *not* include bindings. Those will be included when we
deprecate the legacy API.