Manage collision geometries with DeformableRigidManager
- Add RegisterCollisionObjects() that registers all rigid collision
geometries registered in MultibodyPlant at the corresponding
DeformableRigidManager.
- Add RegisterDeformableGeometries() that registers all deformable
geometries from DeformableModel at the corresponding
DeformableRigidManager.
- Add UpdateCollisionOjbectPoses() and UpdateDeformableVertexPositions()
that update the geometries based on context.

This is meant to be a temporary solution while SceneGraph doesn't
support deformable geometries yet. When SceneGraph takes control of
deformable geometries, much of the code here can be migrated to
SceneGraph with some modifications.