parsing: Add support for SDFormat 1.8 model composition (redo) (#15099)
This a redo of #14401 but does not create a frame for each model in the parent scope of the model. There is still a frame associated with each model, but it's named __model__ and is found inside the model instance.

Changes in behavior:
* Models nested via the `<include>` tag are no longer flattened, which
  means each included model will have its own model instance.
* `drake::multibody::Parser::AddAllModelsFromFile` now returns all added
  models including nested models

Note: This does not use a custom parser for URDFs via libsdformat's
Interface API, and thus may incur unexpected behavior when including
URDF files from SDFormat pending the full resolution of #14295.

* Do not create frame named after model

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>