Use kDefaultPortName throughout
This is in preparation for deprecating the overloads that allow the user
to omit the port name.

In many cases we should probably be using a more specific string literal
instead of the kDefaultPortName, but for simplicity this commit converts
everything to kDefaultPortName in bulk.  We can go back through each
directory later one by one and have the subject matter experts choose
better names.

The only call sites not changed are the direct unit tests for System and
LeafSytem (and their Python bindings).  Those need to be more carefully
adjusted as part of the deprecation commit.