[framework] Remove manipulation_station dependency from test (#15196)
* [framework] Remove manipulation_station dependency from test

The system_html_test used ManipulationStation to demonstrate a large and
complex diagram. The problem was that bazel test //systems/framework/...
was therefore rebuilding a huge amount of additional code, and seriously
slowing down development on patches that change the framework. I've been
using a version of this patch locally for months any time I'm working on
systems, but I think its time to improve it for everyone else.