Add Schur complement
Support deformable rigid contact by providing the functionality to
calculate the Schur complement of a symmetric positive definite matrix.
With Schur complement, the deformable dynamics tangent matrix can be made
into a smaller matrix involving only the dofs in contact. The velocity of
the dofs not in contact can also be solved using Schur complement  after
the contact solve.