[geometry] Introduce BvhUpdater and add it to DeformableVolumeMesh (#15221)
The BvhUpdater is a "modifier" for a Bvh. Given a mesh and its
corresponding Bvh, it causes the Bvh to update given the current
position of the mesh's vertices -- it assumes no topological changes.

This initial implementation is simple.
  - It only updates BVHs built on AABB.
  - It only refits bounding volumes -- it doesn't change the topology
    at all.

This also adds a bvh and bvh updater to the DeformableVolumeMesh. This
keeps the deformable mesh in sync.

(Incidentally cleans up some redundant semantics in the MeshDeformer -
removing set_mesh() and documenting why it isn't necessary.)