geometry:optimization Implements IsBounded and Hyperellipsoid Volumes
Also includes a ConvexSets (plural) container as my proposed solution to memory management for workflows like this:
```
ConvexSets MakeIrisObstacles(
    const QueryObject<double>& query_object,
    std::optional<FrameId> reference_frame = std::nullopt);
HPolyhedron Iris(const ConvexSets& obstacles,
                 const Eigen::Ref<const Eigen::VectorXd>& sample,
                 const HPolyhedron& domain,
                 const IrisOptions& options = IrisOptions());
```

Also includes a number of small fix-ups and style updates:
- HyperEllipsoid => Hyperellipsoid (per discussion with Sean)
- AddPointInSetConstraint => AddPointInSetConstraints
- expressed_in => reference_frame
This are API breaking changes, but we marked the classes as Experimental so that we could do this in the early stages.

This is not chopped up quite as finely as I've been trying to do so far.  But it's the last batch of changes, which had been piling up, before I land the IRIS algorithm in the next PR.  I hope it's ok(better?) this way for my reviewers!