Extend copyable_unique_ptr const correctness
In order to be compatble with a future copy-on-write unique pointer,
copyable_unique_ptr made a distinction between get() and get_mutable().
The idea being that it would be nice to swap copyable with copy on write
with the absolute minimum number of changes. So, by declaring that const
distinction now, callers will think about mutable access.

However, this allowed getting a non-const *reference* via dereferencing,
even from a const copyable_unique_ptr. Therefore, we distinguish
derefencing a copyable_unique_ptr based on the *container's* constness.
Dereferencing a const copyable_unique_ptr will *always* return a const
reference (and will be preserved with a copy-on-write pointer). But
dereferencing a non-const container, when swapped to copy-on-write, will
lead to the copy.

This also adds a novel constructor: a model constructor. Constructs a new,
uniquely owned instance by a copying a given *value*.