[common] Clean up ExpressionCell const-ness
The prior change to remove const from the shared_ptr left open the const
propagation hole when directly accessing the member field.

Here, we introduce getters for the cell so that constness is preserved,
and switch the downcasters to return direct cell references (instead of
shared_ptr references) because none of the callers actually need to
indirectly increment the shared_ptr count.