SignalLogger: Remove context-per-thread hazard (breaking)
Relevant to: #10228

Move the internal SignalLog member to become a cache entry. Abandon the data
access methods on SignalLogger, and instead provide ergonomic lookup methods to
get the SignalLog itself.

Make SignalLog copyable and movable for compatibility with the cache.

Provide complete python bindings for SignalLogger and SignalLog.

Update numerous tests and examples.

This patch is breaking; it removes existing (if problematic) APIs without
deprecation.