Approximate signed distance field for deformable bodies (#15300)
* Augment reference configuration deformable geometry with an approximate
signed distance field. We precompute the signed distance at each vertex
of the deformable mesh at *reference* configuration and approximate the
signed distance at the contact point in the *deformed* configuration by
linearly interpolating the precomputed values at the vertices.
* The domain of the approximated signed distance field is strictly in the
interior of the deformable mesh and the value is non-positive.
* Collect this approximated signed distance information with
DeformableContactData.