[geometry/optimization] Add GraphOfConvexSets
This first PR adds the graph construction / maintenance logic only.  I've separated the actual mathematical programming part to follow-up PRs, so that they can be reviewed more carefully.

The additional correctness of the added costs/constraints on an edge will be established in the next PR, once the Solve method becomes available.