Remove TriggerTest from leaf_system_test.cc
This test had lost all meaning. It was predicated on antiquated use of
the event APIs. The suspicion is that its value has gradually eroded since
its initial incarnation. Rather than trying to mutate it into something
that can have meaning, it seemed simpler to remove it whole sale and add
in a new test.