[solvers] Add bindings for MathematicalProgram::AddCost(shared_ptr, vars) (#15400)
There are a number of very similar entry points for adding costs to mathematical program.  We only binding a subset of them so far.  I elected to bind this variant because we don't currently binding the constructor for the Binding_* classes (although I've confirmed that this is also easy to do).

This also follows the pattern in AddConstraint.  The AddConstraint((shared_ptr) variant was already bound, but AddConstraint(Binding<Constraint>) is not (yet).

For most cases, we've avoided needing this because MathematicalProgram provides enough sugar methods to construct and add all of the variants of costs and constraints that user have been asking for.  But L2NormCost is a new case (until we outwait the deprecation period) where we cannot offer the right sugar.