Label spatial force on object in drake visualizer
For visualizing hydroelastic contact, the visualized quantities would
label a pair of geometries and show various quantities. One was labeled
simply "Spatial force" but didn't indicate which body in the pair the
force was acting on. This adds the name of the body being acted on.

For point-pair contact, we also label which body the force is applied to.
There's some additional complexity for point pair contact due to the fact
that there are no guarantees on body ordering. This also adds notes on
that specific subject (and a TODO in MBP to address this).