[systems] Add one elementary benchmark
This first benchmark only measures the throughput of PassThroughSystem
and Diagram when used with vector-valued ports.  Even so, it is enough
to highlight unnecessary spending on dynamic casts.

Factor out a common cc_binary rule between the existing cassie_benchmark
and the newly-added systems/benchmarking, as well as a common cc_library
rule for new statistics, to reduce copy-paste across Drake's benchmarks.