[framework] Never dynamic_cast an AbstractValue (#15419)
The AbstractValue implementation goes to great lengths to avoid dynamic_cast,
because RTTI checking is exceedingly slow, relative to what we expect during
an inner loop. We should use its built-in type checking functions, instead of
casting ourselves.

The leaf_system.cc vector cast was showing up as a profiling hot spot, but
even without that we should make this change, even in test code. Trying to
directly cast sets a bad example for others to imitate.

This provides a 5% speed-up on the PassThrough3 benchmark.