planar visualizer uses SceneGraph API to visualize (breaking)
The planar visualizer had previously used a hackneyed approach to
visualize. It would generate a DrakeVisualizer lcm message for loading
and consume pose bundle for updating poses. This updates it to make use
of the SceneGraphInspector and QueryObject for those tasks (akin to what
meshcat visualizer does).

This introduces a *breaking* change. Input port 0 is no longer PoseBundle-
valued. It is now QueryObject-valued. Thus it should be connected to a
different port on SceneGraph.