[common] Fine tune {not_,}equal_strict<Expression> for Eigen 3.4 (#15453)
In various Eigen algorithms, it uses "short-circuit when zero" guards to
skip computationally expensive steps in cases where knows the end result
will remain unchanged.

When we have Expressions filled with only constants, this works fine.
However, when we have Expressions filled with Variables, trying to check
for zero throws an exception because the resulting Formula cannot convert
to back to a bool -- and in any case, we want the full expression tree
anyway, without any short-circuits, for doing symbolic evaluation later.

Therefore, we tweak Eigen::numext::equal_strict and not_equal_strict to
special-case their results if either of the operands is a literal zero,
with no throwing even if the other operand has unbound variables.