[common] Remove copy/move from ExpressionCell
We do not appear to take advantage of the copy / move construction at all
(and we have const members, so I suspect move was doing copies anyway).

For optimizations like opportunistic mutation or more careful sharing,
it seems like having implicit copies available might be a bit of a hazard.

Similarly, the base class default destructor is dead code, so is removed.

Fix a stray comment on a factory default constructor.