Add and register top-level Drake package.xml
This change (re-)adds a top-level package manifest for Drake, and
registers it in all PackageMap instances during construction. This
should enable URI resolution that currently utilizes individual package
manifests within the Drake source tree to switch to the new top-level
manifest with relatively little effort.

Note that the new package manifest conforms with the appropriate schema,
but contains no dependency information and has a placeholder version
number (0.0.0) in place of the actual Drake version. These are not
necessary to use the manifest for resolving URIs.

Co-authored-by: Scott K Logan <logans@cottsay.net>