primitives: Introduce VectorLogSink
Relevant to: #10228

This is step 2 of 4 to deprecate and replace SignalLogger and SignalLog
with a logging system that avoids threading hazards (SignalLogger) and
slow memory management (SignalLog).

VectorLogSink will eventually replace SignalLogger.  The chief
improvement is relocation of log storage from member data in a System(!)
to storage per-context via a scratch cache entry. The direct-to-log
accessors and mutators are removed, but access is clarified and still
convenient via {Get,Find}[Mutable]Log methods.