[geometry] Add declaration to *remove* collision filters (#15415)
Previous API only allowed for *adding* filters on geometry pairs. This
introduces the API that can allow (generally) the removal of a filter from
a geometry pair.

This introduces the concept of the "permanent" filter. It supports scene
graph invariants (e.g., geometries affixed to the same frame are always
filtered, etc.) This concept is purely internal -- it doesn't affect the
public API.

The code changes include:

  - CollisionFilterDeclaration gets new statement methods.
  - CollisionFilter learns how to apply them.
  - CollisionFilterManager documents the "invariant" filters.
  - Touch numerous tests that depend on having a collision filter.
  - Python bindings for new API.