[visualization] planar visualizer uses SceneGraph API to visualize (#15456)
The planar visualizer had previously used a hackneyed approach to
visualize. It would generate a DrakeVisualizer lcm message for loading
and consume pose bundle for updating poses. This updates it to make
use of the SceneGraphInspector and QueryObject for those tasks (akin
to what meshcat visualizer does).

This is done in a deprecated, backwards-compatible way. The old
spelling will still work, but if the visualizer is connected to a
PoseBundle-valued port, the first call to draw() will dispatch a
deprecation warning.

This lead to the following ancillary changes:
  - All uses of PlanarSceneGraphVisualizer in Drake have been updated
    to use the new API (in fact, they've been updated to prefer using
    ConnectPlanarSceneGraphVisualizer).
  - The old unit tests have simply been copied wholesale. One test
    fixture is based on the new implementation, the other, labeled
    "Deprecated" on the old.
    - This also led to a change in the test for ConnectPlanar...
      The old test called diagram.Publish(context) on a diagram
      containing visualizers. However, those visualizers had their
      _shown flag set to False, so no drawing work was actually being
      done. The test has been reformulated to really test the drawing
      logic.