[geometry/optimization] Adds IrisInConfigurationSpace (#15476)
* [geometry/optimization] Adds IrisInConfigurationSpace

This is a first pass at the algorithm, using Snopt/Ipopt but not (yet) backed by Ibex/dReal.  That will come in the next PR.

There is still an open question about whether we can use ONLY Ibex/dReal, instead of Snopt.  But based on python experiments, I believe we can use Snopt first and then follow with Ibex/dReal to confirm that our sets are collision free. Furthermore, this Snopt version is useful immediately, and including Ibex/dReal will not change the interface, only the backend.