Deprecate SceneGraphInspector::all_frame_ids()
The old API was leaking internal:: classes into the public API.
The `GeometryState::FrameIdRange` is an alias for
internal::MapKeyRange which, itself, is templated on
internal::InternalFrame -- so doubly sinful. And when we push
GeometryState into internal:: it'll be three sins. This deprecates
the API and makes it match GetAllGeometryids (in terms of behavior
and performance).

This required changes to the MapKeyRange itself in order to
instantiate vector(range.begin(), range.end()) -- it supports the
std::iterator_traits directly.

Also deprecates the bindings, binds the new API an updates calls to
the old API to the new API.