[WIP] Concept demonstration of ProceduralSystem
Here is one concept of writing and debugging procedrual code in the systems framework.  I also think we should offer FiniteStateMachine isa LeafSystem and/or FinieStateMachine isa Diagram, but this, I think, it what is closest in spirit to some discussions.

This example (of a quadrotor) was chosen to illustrate a high-rate control task + low-rate procedural code living in harmony.

There are some very rough edges (it's been a while since I've thought hard about multithreading!), but I hope we can discuss the concept and it's possibilities/limitations before we tear apart my exception vulnerability.