[common] Deprecate DRAKE_THROW_UNLESS on a bare pointer
For a raw pointer (e.g., 'foo' in 'void DoStuff(Foo* foo);'), when checking
for null we'd like to insist on an explicit mention of nullptr:

Correct: DRAKE_THROW_UNLESS(foo != nullptr);
Incorrect: DRAKE_THROW_UNLESS(foo);

Because the macro is intended to provide a useful exception message to users,
we should err on the side of extra detail about the failure condition.  The
meaning of "foo" isolated within error message text does not make it clear
that a null pointer is the proximate cause of the problem.