[solvers] Fix assertion in AddPositiveSemidefiniteConstraint
The symbolic transpose assertion was being incorrectly skipped, so we must
rewrite the check to use CheckStructualEquality instead of operator==.

The CheckStructualEquality doesn't work well with MatrixBase<Derived>, so we
must rewrite the function to take a Ref, instead. (This is an improvement in
any case; we over-use Derived in the MathematicalProgram API.)

Also of note, with the Eigen 3.4 pre-release, the assertion was actually being
compiled down to a bool, rather than skipped, leading to a type check error.